📝 docs(library.py): add function descriptions This commit adds descriptions to the functions in the library.py file to improve code readability and maintainability. The descriptions are as follows: - create_repo: creates a new repository in Github - get_list_of_branches: gets a list of the TFS repo branches - list_files: lists all files in a directory and writes the output to a file - getting_binary_extensions: classifies binary extensions - upload_binary_to_git_lfs: uploads binary files to Git Large File Storage in Github - file_with_extension: gets files with a specific extension.
📝 docs(library.py): add docstring to clone_target_git function
The `clone_target_git` function now has a docstring that describes its purpose, which is to clone the GitHub repository. This improves the readability and maintainability of the code.

📝 docs(main.py): add docstring to migration function
🔥 refactor(main.py): remove unnecessary line break
The migration function now has a docstring to improve code readability and maintainability. The line break after the shutil.rmtree function call has been removed as it is unnecessary.