📝 docs(README.md): add missing code block Added a missing code block to the README.md file in both the Discovery_script_dir and Migration_demo_dir directories.
📝 docs(README.md): Add flow chart and code file details, update pre-requisites and goals
This commit adds a flow chart to the README.md file that shows the migration process from TFS to GitHub and ADO. It also adds details about the code files, including main.py, library.py, and credentials.py. The pre-requisites section has been updated to include Python 3.9 or greater, the requirements.txt file, Github credentials, and git tfs. The goals section has been updated to include the migration of user and access control to GitHUB and ADO. Additionally, the README.md file has been updated to reflect the changes in the steps involved in migrating source code, work items, and user privileges from TFS to GitHub & ADO.

📝 docs(README.md): add missing newline character
A newline character was added to the end of the file to ensure proper formatting of the file.